refactor: create commponet folder
