+++ frontend/src/components/Accessory/EquipmentAccessoryRow.jsx
