+++ frontend/src/components/Accessory/EquipmentCard.jsx
